RAG/requirements.txt
llama-index-core>=0.9.0
llama-index-llms-groq>=0.1.0
llama-index-vector-stores-pinecone>=0.1.0
llama-index-embeddings-cloudflare>=0.1.0
llama-index-postprocessor-cohere-rerank>=0.1.0
llama-parse>=0.1.0
tavily-python>=0.2.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
tenacity>=8.2.0
numpy>=1.24.0
pydantic>=2.0.0
pinecone-client>=3.0.0
cohere>=4.0.0
